clean up stm32f103 mcu def
