+++ electronics_lib/__init__.py
@@ -29,7 +29,6 @@ from .PowerConditioning import BufferedSupply, Supercap, SingleDiodePowerMerge, 
 
 from .Microcontroller_Lpc1549 import Lpc1549_48, Lpc1549_64
 from .Microcontroller_Stm32f103 import Stm32f103_48
-from .Microcontroller_Stm32f103_new import Stm32f103_48_new
 from .Microcontroller_Nucleo32 import Nucleo_F303k8
 from .Microcontroller_nRF52840 import Holyiot_18010_Nrf52840
